[NIFI-9666] Sepaartion of C2 Client(s) and the creation of the heartbeats. Pushing as much as possible under C2.
