+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/pom.xml
@@ -177,12 +177,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>c2-client-base</artifactId>
-            <version>1.17.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>c2-client-http</artifactId>
+            <artifactId>c2-client-service</artifactId>
             <version>1.17.0-SNAPSHOT</version>
         </dependency>
         <dependency>
